backend/requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-cors==3.0.10
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytest==7.3.1
pytest-mock==3.10.0
uvicorn==0.21.1
psutil==5.9.5 